Added json as static library
