+++ src/engine/source/base/baseTypes.hpp
@@ -1,7 +1,7 @@
 #ifndef _H_BASE_TYPES
 #define _H_BASE_TYPES
 
-#include "json.hpp"
+#include <json/json.hpp>
 #include "result.hpp"
 
 namespace base
