+++ src/engine/source/builder/asset.hpp
@@ -8,7 +8,7 @@
 
 #include "definitions.hpp"
 #include "expression.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 #include "registry.hpp"
 
 namespace builder
