+++ src/engine/source/builder/builders/operationBuilder.cpp
@@ -4,7 +4,7 @@
 
 #include "baseTypes.hpp"
 #include "expression.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 #include "registry.hpp"
 #include "result.hpp"
 #include "syntax.hpp"
