+++ src/engine/source/builder/builders/stageBuilderOutputs.cpp
@@ -9,7 +9,7 @@
 #include <logging/logging.hpp>
 
 #include "expression.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 #include "registry.hpp"
 
 namespace builder::internals::builders
