+++ src/engine/source/builder/builders/stageBuilderParse.cpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "expression.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 #include "registry.hpp"
 
 namespace builder::internals::builders
