+++ src/engine/source/builder/environment.hpp
@@ -12,7 +12,7 @@
 #include "asset.hpp"
 #include "expression.hpp"
 #include "graph.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 #include "registry.hpp"
 
 namespace builder
