+++ src/engine/test/source/builder/builder_test.hpp
@@ -4,7 +4,7 @@
 
 #include "connectable.hpp"
 #include "rxcpp/rx.hpp"
-#include "json/json.hpp"
+#include <json/json.hpp>
 #include <hlp/hlp.hpp>
 
 std::map<std::string, std::string> decoders = {
