+++ src/engine/test/source/builder/builders/operationBuilder_test.cpp
@@ -2,7 +2,7 @@
 
 #include "baseTypes.hpp"
 #include "builder/builders/operationBuilder.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 
 using namespace builder::internals;
 using namespace builder::internals::builders;
