+++ src/engine/test/source/builder/builders/stageBuilderMap_test.cpp
@@ -4,7 +4,7 @@
 #include "builder/builders/operationBuilder.hpp"
 #include "builder/builders/stageBuilderMap.hpp"
 #include "builder/registry.hpp"
-#include "json.hpp"
+#include <json/json.hpp>
 
 using namespace builder::internals;
 using namespace builder::internals::builders;
